Refactor web v2 app shell
